Pause Fedora jobs
Per the linked bugs, this does not seem to have a quick resolution.
The bug mentions that the copr nightly repos have other known issues,
so there just does not seem to be a good solution until fixed packages
are imported upstream.  For now, pause Fedora jobs to get the gate
working.

Change-Id: Ib349592c3fcf2008fcab9ae8875de1c997eef7d5