bindep: remove lsb-release
This came in with edc06a20e57bdf0a9a03949b54395008d73520dc although
it's not totally clear why.

We found that we were sourcing devstack scripts during build in infra,
and inadvertently relying on this package being around on the host.
This was incorrect but can this type of problem is hidden if the
package is installed unnecessarily.

I do not think this actually required now; remove from bindep to
minimise chance for such false or hidden dependencies.

Change-Id: I7895d4c5019511eb2169915cd8092e707e8ee732
Story: #2007407
Task: #39068