Switch functional tests to containers
This switches the functional tests to container based build jobs.
This is mostly transparent for the existing build and boot jobs.

The release test is a bit interesting here.  The nodepool
container-based release job doesn't work in this context because it
doesn't have a way to also include dib from the zuul checkout into the
container -- ergo it's pointless to run in the dib gate.

We should probably reorganise the nodepool siblings job so it can
check some projects out from tags, then we could run something similar
to this with nodepool/openstacksdk/glean etc checked out from tags if
we want.

We can handle this in a future update so we can test these images
which we want to build in our production container asap.

Depends-On: https://review.opendev.org/722556
Depends-On: https://review.opendev.org/722549
Change-Id: Ieaf07438c145ba609d4d7b9a055e91b894eca6d9