Install yum and yum-utils on Red Hat family hosts
Elements like yum and yum-minimal depend on the yum and yumdownloader
CLIs on the DIB host. We have been assuming these tools are always
installed on Red Hat family distributions but that is no longer the case
starting from RHEL 8 and CentOS 8 and Fedora for a while -- these
systems moved on to DNF and DNF does not have a hard requirement on yum.

We noticed these two tools were missing while building an Octavia
amphora image on a CentOS 8 controller node in CI:

[...] 08-yum-chroot:_install_pkg_manager:223  : sudo -E yum -y [...]
[...] sudo: yum: command not found

Change-Id: I65807751b3764e86ba5adeb47552305ef2ea299b