Correctly create DIB_ENV variable and dib_environment file
The DIB_ENV variable can contain multiline content, for example at
least DIB_BLOCK_DEVICE_CONFIG is a multiline YAML data [1], so for
this type of content the variable created with incomplete data, also
this incomple data echoed [2] when creating the dib_environment file
which creates an unusable file. This change fixes the issue.

1. https://github.com/openstack/diskimage-builder/blob/79ea63f5251f6f9a11d1a26537cd2c76389d2d66/doc/source/user_guide/building_an_image.rst#disk-image-layout
2. https://github.com/openstack/diskimage-builder/blob/79ea63f5251f6f9a11d1a26537cd2c76389d2d66/diskimage_builder/elements/manifests/cleanup.d/01-copy-manifests-dir#L32

Change-Id: I3b74ede69eb064ad813a9108ec68a228e549e8bb
Signed-off-by: Maksim Malchuk <maksim.malchuk@gmail.com>